commit 0ca566afef19b097b12a7218eb3cb4de37897162
Author: Abhishek Bagusetty <abagusetty@login2.crusher.olcf.ornl.gov>
Date:   Fri Apr 8 15:50:56 2022 -0400

    [HIP] fixed numerically inconsistent results for (T) kernel

commit 3c507695f48c058c78316e5ee591a89083f1f3d2
Author: Abhishek Bagusetty <abagusetty@anl.gov>
Date:   Fri Apr 8 17:47:12 2022 +0000

    [CUDA] remove const-ness for GPU streams

commit 4cac786f46fc0d4d2d3a62c2ad93c96713b4cc62
Author: Abhishek Bagusetty <abagusetty@anl.gov>
Date:   Fri Apr 8 16:42:43 2022 +0000

    [CUDA][HIP][SYCL] gpu stream pool cleanup

commit b4ea9289a79059e85cdfacef7e6ef170943d9b01
Author: Abhishek Bagusetty <abagusetty@anl.gov>
Date:   Thu Apr 7 02:28:41 2022 +0000

    formatting cleanup

commit c20c4f8d3037214a3d5a8075475c6daf9efa46a5
Author: Abhishek Bagusetty <abagusetty@anl.gov>
Date:   Wed Apr 6 23:34:40 2022 +0000

    [CUDA][HIP][SYCL] pass a gpuStream by ref rather than pointer to GPU kernels

commit e3762126683a0becff0b30a38efbccf9465d2eef
Merge: 3793319c 103e21a3
Author: Abhishek Bagusetty <abagusetty@anl.gov>
Date:   Wed Apr 6 23:23:05 2022 +0000

    Merge branch 'gpu_stream_pool' of github.com:NWChemEx-Project/TAMM into gpu_stream_pool

commit 3793319c6bea7b61883c7a54ca0b1ce0d2c65b24
Author: Abhishek Bagusetty <abagusetty@anl.gov>
Date:   Wed Apr 6 23:21:49 2022 +0000

    [CUDA][HIP][SYCL] create a singleton streampool

commit 103e21a38d49743ecc3dc0e0a168c5330bfab222
Author: Ajay Panyala <ajay.panyala@gmail.com>
Date:   Mon Apr 4 22:28:39 2022 -0700

    dump cmakebuild

commit 36975df3056a2ecf31709b3f74c21e28b50791f7
Author: Abhishek Bagusetty <abagusetty@login2.crusher.olcf.ornl.gov>
Date:   Mon Apr 4 22:15:39 2022 -0400

    [HIP] hipLaunchHostFunc not yet supported

commit 19f90f29e9259badf59f3531ed6cb32132c1daad
Author: Abhishek Bagusetty <abagusetty@anl.gov>
Date:   Tue Apr 5 00:35:06 2022 +0000

    [SYCL][CUDA][HIP] fix the gpu steram pool to use std::array of unique_ptrs

commit d27c1adba957a69f9ac6dfea33cd9556e1cfe219
Merge: e7588c2e 09cf2fb0
Author: Abhishek Bagusetty <abagusetty@anl.gov>
Date:   Mon Apr 4 17:19:41 2022 +0000

    Merge branch 'develop' into gpu_stream_pool

commit e7588c2e92dbc61b45e1c250de546c3bc238584e
Author: Abhishek Bagusetty <abagusetty@anl.gov>
Date:   Mon Apr 4 01:06:55 2022 +0000

    [SYCL] make queue in-order

commit 9abaaab9c1a764c59eaf3727f34a7534e3df664c
Merge: 6f65d5cc 82719bdd
Author: Abhishek Bagusetty <abagusetty@anl.gov>
Date:   Sat Apr 2 18:40:25 2022 +0000

    Merge branch 'gpu_stream_pool' of github.com:NWChemEx-Project/TAMM into gpu_stream_pool

commit 6f65d5cc1f3757b05ecc4a91baf1bba7749cca03
Author: Abhishek Bagusetty <abagusetty@anl.gov>
Date:   Sat Apr 2 18:19:20 2022 +0000

    [SYCL][CUDA][HIP] fix speedup for (T) kernel

commit 82719bdde959513cb8719d7931739ed64e939d73
Author: Ajay Panyala <ajay.panyala@gmail.com>
Date:   Thu Mar 24 07:44:08 2022 -0700

    update compare script

commit 4f9a1923482520e15e801055c62608486ba37b59
Merge: 8965214f 0a774293
Author: Abhishek Bagusetty <abagusetty@anl.gov>
Date:   Wed Mar 23 14:11:26 2022 +0000

    Merge branch 'develop' into gpu_stream_pool

commit 8965214f19a081a746da2b7a821a281ec0109a09
Author: Ajay Panyala <ajay.panyala@gmail.com>
Date:   Tue Mar 22 21:18:46 2022 -0700

    (T): reduce mem. req. for computing full V2

commit e9c26f37ef409d77ba44a7d6ad51fe1924468243
Merge: 564ac984 f55fbba8
Author: Abhishek Bagusetty <abagusetty@anl.gov>
Date:   Wed Mar 23 02:54:43 2022 +0000

    Merge branch 'gpu_stream_pool' of github.com:NWChemEx-Project/TAMM into gpu_stream_pool

commit 564ac984899b732513ebde5a0cdc8aaffa27881d
Author: Abhishek Bagusetty <abagusetty@anl.gov>
Date:   Wed Mar 23 02:54:31 2022 +0000

    [SYCL, HIP, CUDA] overlap copy-comp, streams cleanup, events cleanup

commit f55fbba8dc187026e152ba7e48bc5887763ff042
Merge: ff3e53d5 0f2c2ae5
Author: Ajay Panyala <ajay.panyala@gmail.com>
Date:   Tue Mar 22 18:40:45 2022 -0700

    Merge branch 'develop' into gpu_stream_pool

commit ff3e53d59e30146e1adfc9335cc3dd89c217c0dc
Merge: 213c95b2 c44fd2fe
Author: Abhishek Bagusetty <abagusetty@anl.gov>
Date:   Wed Mar 16 03:09:22 2022 +0000

    [SYCL] address merge conflicts with develop

commit 213c95b2537e6a6388e546140fbec1396e64d562
Merge: 428ea736 5e6d7030
Author: Ajay Panyala <ajay.panyala@gmail.com>
Date:   Sun Mar 13 22:51:47 2022 -0700

    merge develop

commit 428ea736daf9924173f4fb589a1b5660e31c682f
Author: Ajay Panyala <ajay.panyala@gmail.com>
Date:   Thu Mar 10 15:40:01 2022 -0800

    fix

commit 66a5aa40fb881c997e2b187613eb8c9fc133f759
Author: Abhishek Bagusetty <abagusetty@anl.gov>
Date:   Tue Mar 8 06:26:49 2022 +0000

    [SYCL] fix a bug in (T) kernel leading to NaNs

commit c6d8f47d9aed7729dc806a67424aba6b30f06f19
Author: Abhishek Bagusetty <abagusetty@anl.gov>
Date:   Tue Mar 8 02:48:42 2022 +0000

    gou streams, cleanup

+++ methods/scf/scf_taskmap.hpp
@@ -57,7 +57,7 @@ void readLoads(std::vector<NODE_T> &s1_all, std::vector<NODE_T> &s2_all,std::vec
     NODE_T s2;
     VAL_T nTasks;
     
-    for(int i=0;i<s1_all.size();i++)    
+    for(size_t i=0;i<s1_all.size();i++)
     {
         s1 = s1_all[i];
         s2 = s2_all[i];
